♻️ refactor: Refactor hook
